Modularize lobby game search panel
